fixed shiftReviews, start on shift status

+++ server/routes/shiftReviews.js
@@ -11,27 +11,37 @@ module.exports = function shiftReviewRoutes(app) {
 
   // user reviews TODO validate user ID
   .post('/user/:userID/shift/:shiftID', function* createUserReview() {
+    const user = this.passport.user;
     const userID = this.params.userID;
-    const shiftID = this.params.shiftID;
-    const requestJson = this.request.body;
-    requestJson.fk_ShiftReviewOnUser_userID = userID;
-    requestJson.fk_ShiftReviewOnUser_shiftID = shiftID;
-    const q = {};
-    q.sql = 'INSERT INTO ?? SET ?';
-    q.values = ['ShiftReviewOnUser', requestJson];
-    // console.log(q);
-    this.body = yield query(q);
+    if (user.scope.userID.toString() === userID) {
+      const shiftID = this.params.shiftID;
+      const requestJson = this.request.body;
+      requestJson.fk_ShiftReviewOnUser_userID = userID;
+      requestJson.fk_ShiftReviewOnUser_shiftID = shiftID;
+      const q = {};
+      q.sql = 'INSERT INTO ?? SET ?';
+      q.values = ['ShiftReviewOnUser', requestJson];
+      // console.log(q);
+      this.body = yield query(q);
+    } else {
+      this.body = {message: 'no permission'};
+    }
   })
 
   // TODO validate userID
   .put('/user/:userID/shift/:shiftID', function* updateUserReview() {
+    const user = this.passport.user;
     const userID = this.params.userID;
-    const shiftID = this.params.shiftID;
-    const requestJson = this.request.body;
-    const q = {};
-    q.sql = 'UPDATE ?? SET ? WHERE ?? = ? AND ?? = ?';
-    q.values = ['ShiftReviewOnUser', requestJson, 'fk_ShiftReviewOnUser_userID', userID, 'fk_ShiftReviewOnUser_shiftID', shiftID];
-    this.body = yield query(q);
+    if (user.scope.userID.toString() === userID) {
+      const shiftID = this.params.shiftID;
+      const requestJson = this.request.body;
+      const q = {};
+      q.sql = 'UPDATE ?? SET ? WHERE ?? = ? AND ?? = ?';
+      q.values = ['ShiftReviewOnUser', requestJson, 'fk_ShiftReviewOnUser_userID', userID, 'fk_ShiftReviewOnUser_shiftID', shiftID];
+      this.body = yield query(q);
+    } else {
+      this.body = {message: 'no permission'};
+    }
   })
 
   .get('/user/:userID/shift/:shiftID', function* getUserReview() {
@@ -53,36 +63,51 @@ module.exports = function shiftReviewRoutes(app) {
 
   // TODO validate userID
   .delete('/user/:userID/shift/:shiftID', function* deleteUserReview() {
+    const user = this.passport.user;
     const userID = this.params.userID;
-    const shiftID = this.params.shiftID;
-    const q = {};
-    q.sql = 'DELETE FROM ?? WHERE ?? = ? AND ?? = ?';
-    q.values = ['ShiftReviewOnUser', 'fk_ShiftReviewOnUser_userID', userID, 'fk_ShiftReviewOnUser_shiftID', shiftID];
-    this.body = yield query(q);
+    if (user.scope.userID.toString() === userID) {
+      const shiftID = this.params.shiftID;
+      const q = {};
+      q.sql = 'DELETE FROM ?? WHERE ?? = ? AND ?? = ?';
+      q.values = ['ShiftReviewOnUser', 'fk_ShiftReviewOnUser_userID', userID, 'fk_ShiftReviewOnUser_shiftID', shiftID];
+      this.body = yield query(q);
+    } else {
+      this.body = {message: 'no permission'};
+    }
   })
 
   // facility reviews TODO validate facility ID
   .post('/facility/:facilityID/shift/:shiftID', function* createFacilityReview() {
+    const user = this.passport.user;
     const facilityID = this.params.facilityID;
-    const shiftID = this.params.shiftID;
-    const requestJson = this.request.body;
-    requestJson.fk_ShiftReviewOnFacility_facilityID = facilityID;
-    requestJson.fk_ShiftReviewOnFacility_shiftID = shiftID;
-    const q = {};
-    q.sql = 'INSERT INTO ?? SET ?';
-    q.values = ['ShiftReviewOnFacility', requestJson];
-    this.body = yield query(q);
+    if (user.scope.facilityID.toString() === facilityID) {
+      const shiftID = this.params.shiftID;
+      const requestJson = this.request.body;
+      requestJson.fk_ShiftReviewOnFacility_facilityID = facilityID;
+      requestJson.fk_ShiftReviewOnFacility_shiftID = shiftID;
+      const q = {};
+      q.sql = 'INSERT INTO ?? SET ?';
+      q.values = ['ShiftReviewOnFacility', requestJson];
+      this.body = yield query(q);
+    } else {
+      this.body = {message: 'no permission'};
+    }
   })
 
   // TODO validate facility ID
   .put('/facility/:facilityID/shift/:shiftID', function* updateFacilityReview() {
+    const user = this.passport.user;
     const facilityID = this.params.facilityID;
-    const shiftID = this.params.shiftID;
-    const requestJson = this.request.body;
-    const q = {};
-    q.sql = 'UPDATE ?? SET ? WHERE ?? = ? AND ?? = ?';
-    q.values = ['ShiftReviewOnFacility', requestJson, 'fk_ShiftReviewOnFacility_facilityID', facilityID, 'fk_ShiftReviewOnFacility_shiftID', shiftID];
-    this.body = yield query(q);
+    if (user.scope.facilityID.toString() === facilityID) {
+      const shiftID = this.params.shiftID;
+      const requestJson = this.request.body;
+      const q = {};
+      q.sql = 'UPDATE ?? SET ? WHERE ?? = ? AND ?? = ?';
+      q.values = ['ShiftReviewOnFacility', requestJson, 'fk_ShiftReviewOnFacility_facilityID', facilityID, 'fk_ShiftReviewOnFacility_shiftID', shiftID];
+      this.body = yield query(q);
+    } else {
+      this.body = {message: 'no permission'};
+    }
   })
 
   .get('/facility/:facilityID/shift/:shiftID', function* getFacilityReview() {
@@ -105,12 +130,17 @@ module.exports = function shiftReviewRoutes(app) {
 
   // TODO validate facility ID
   .delete('/facility/:facilityID/shift/:shiftID', function* deleteFacilityReview() {
+    const user = this.passport.user;
     const facilityID = this.params.facilityID;
-    const shiftID = this.params.shiftID;
-    const q = {};
-    q.sql = 'DELETE FROM ?? WHERE ?? = ? AND ?? = ?';
-    q.values = ['ShiftReviewOnFacility', 'fk_ShiftReviewOnFacility_facilityID', facilityID, 'fk_ShiftReviewOnFacility_shiftID', shiftID];
-    this.body = yield query(q);
+    if (user.scope.facilityID.toString() === facilityID) {
+      const shiftID = this.params.shiftID;
+      const q = {};
+      q.sql = 'DELETE FROM ?? WHERE ?? = ? AND ?? = ?';
+      q.values = ['ShiftReviewOnFacility', 'fk_ShiftReviewOnFacility_facilityID', facilityID, 'fk_ShiftReviewOnFacility_shiftID', shiftID];
+      this.body = yield query(q);
+    } else {
+      this.body = {message: 'no permission'};
+    }
   });
 
   app.use(shiftReview.routes())
